Exploring more PWR metrics
Been looking at other portfolios, they seem to have a worrisome downward
trend in their PWR values over time.
Thinking to measure the slope and maybe the StdDev of the PWR values.

Want to find a portfolio that is very steady and reliable over time.

Was surprised to see the average PWR value for GB was over 5%!
So the minPWR is actually a bit of an outlier... interesting,
seems the expected PWR could likely be higher than that min.