Planning to regenerate "better than GB" portfolios
Writing results directly to SQLite files.

Probably doing some initial filtering of the assets.
Want to find a "totally different" portfolio than the Golden Butterfly.
See if there are any alternative approaches that could serve as a
complement to GB.

So filter out any assets that are in GB.
And perhaps also filter out most bond assets; people are saying bonds
are at the end of a 40-year bull run. Curious if there are any
portfolios that work well enough even without bonds.