Better comparison of STT and TIPS; not as different
TIPS is still probably a bit better, but the different isn't as dramatic
when you test them over matching time periods.

But STT may be better to use for analysis, since there's 53 years of
data instead of just 37.
That will temper your results, as all the performance numbers are higher
when you look at just 37 years. (Like for GB: 5% PWR instead of 4% PWR).